fix service termination on some chinese OEM devices
some chinese OEM devices have aggressive battery optimizations which
would kill the service when the overlay option is enabled and the
application is swiped away from the recent apps list or after some time
has passed, the fix is a litte bit crude for now. the app saves the last
remaining timeout to the device's storage every one second and then when
the service is killed, it retreives the last saved remaining timeout and
then restarts the service.

This may put a strain on the device's storage but can be avoided if the
overlay is disabled while running on these devices

Change-Id: I7dc0b984499d0e23d1fd5e80b642d7b5c93d9634
Signed-off-by: AbdAlMoniem AlHifnawy <hifnawy_moniem@hotmail.com>